requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
attrs==25.3.0
boto3==1.39.4
botocore==1.39.4
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
dill==0.4.0
docker==7.1.0
fastapi==0.116.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
idna==3.10
importlib-metadata==6.11.0
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.18
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pandas==2.3.1
pathos==0.3.4
platformdirs==4.3.8
pox==0.3.6
ppft==1.7.7
protobuf==5.29.5
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
s3transfer==0.13.0
sagemaker==2.247.1
sagemaker-core==1.0.42
Sastrawi==1.0.1
schema==0.7.7
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
starlette==0.47.1
tblib==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
zipp==3.23.0
